https://trello.com/c/rmiy8Jkx Implement curve interpolation Mistakes in curve interpolation implementation were corrected.

+++ JavaCurveInterpolation/src/main/java/com/ran/dissertation/algebraic/function/DoubleMultifunction.java
@@ -35,7 +35,7 @@ public class DoubleMultifunction<T extends AlgebraicObject<T>> extends DoubleFun
         return new Iterator<T>() {
             private final Iterator<DoubleFunction<T>> doubleFunctionsIterator = doubleFunctions.iterator();
             private DoubleFunction<T> currentFunction = null;
-            private double nextIndex = 0;
+            private int nextIndex = 0;
             @Override
             public boolean hasNext() {
                 return nextIndex <= segments;
